+++ packages/wiz-ui-react/src/components/base/graphs/index.ts
@@ -1,4 +1,5 @@
 export * from "./meter";
+export * from "./line";
 export * from "./transition";
 export * from "./compare";
 export * from "./circle";