+++ packages/styles/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @wizleap-inc/wiz-ui-styles
 
+## 0.9.2
+
+### Patch Changes
+
+- [#818](https://github.com/Wizleap-Inc/wiz-ui/pull/818) [`6fff2934`](https://github.com/Wizleap-Inc/wiz-ui/commit/6fff293452fa178a0ca12d8016cfd2d6eff34e83) Thanks [@okakyo](https://github.com/okakyo)! - [#797] Message Box の icon の幅が変わってしまうバグの修正
+
 ## 0.9.1
 
 ### Patch Changes